Refactor Material and Pass components
