+++ src/Rendering/Components/Material.h
@@ -8,9 +8,31 @@
 
 namespace gore::renderer
 {
-class Pass;
-
 using namespace gore::gfx;
+
+// FIXME: This is a temporary solution to use fixed pass names
+struct PassID
+{
+    const char* k_ShadowPassName = "ShadowPass";
+    const char* k_ForwardPassName = "ForwardPass";
+    const char* k_GBufferPassName = "GBufferPass";
+};
+
+struct Pass
+{
+    GraphicsPipelineHandle shader = {};
+    BindGroupHandle bindGroup[3] = {};
+
+    inline bool operator==(const Pass& other) const
+    {
+        return shader == other.shader && bindGroup[0] == other.bindGroup[0] && bindGroup[1] == other.bindGroup[1] && bindGroup[2] == other.bindGroup[2];
+    }
+
+    inline bool operator!=(const Pass& other) const
+    {
+        return !(*this == other);
+    }
+};
 
 ENGINE_CLASS(Material) final
 {
@@ -22,7 +44,10 @@ public:
 
     void AddPass(const Pass& pass);
 
-    [[nodiscard]] const std::vector<Pass>& GetPasses() const;
+    [[nodiscard]] const std::vector<Pass>& GetPasses() const
+    {
+        return m_Passes;
+    }
 
 private:
     std::vector<Pass> m_Passes;
